App.jsx가 Counter의 상태를 관리하게하라
